:art: [fix] 카카오 SDK script 커스텀 훅으로 빼내기
fix
