+++ src/components/Modal/ShareFolder.jsx
@@ -1,4 +1,3 @@
-import { useEffect } from 'react';
 import styled from 'styled-components';
 
 import ModalTitle from './ModalContent/ModalTitle';
@@ -12,46 +11,44 @@ import modalLink from '../../assets/images/modal_link.svg';
 
 import { copySharingLinkToClipBoard } from '../../utils/copySharingLinkToClipBoard';
 import shareToFacebook from '../../utils/shareToFacebook';
+import useGetKakaoSdkScript from '../../hooks/useGetKakaoSdkScript';
 
 const ShareFolder = ({ modal, onCloseModalButtonClick }) => {
-  useEffect(() => {
-    const script = document.createElement('script');
-    script.src = 'https://developers.kakao.com/sdk/js/kakao.js';
-    script.async = true;
-    document.body.appendChild(script);
-    return () => document.body.removeChild(script);
-  }, []);
+  // useGetKakaoSdkScript 훅으로 카카오 SDK 스크립트를 불러온다
+  const { KakaoSdk } = useGetKakaoSdkScript();
 
   const handleShareToKakao = () => {
-    // url이 id값에 따라 변경되기 때문에 route를 인자값으로 받아줌
-    if (window.Kakao) {
-      const kakao = window.Kakao;
-      if (!kakao.isInitialized()) {
-        kakao.init(process.env.REACT_APP_KAKAO_JS_KEY); // 카카오에서 제공받은 javascript key를 넣어줌
-      }
+    if (KakaoSdk) {
+      // url이 id값에 따라 변경되기 때문에 route를 인자값으로 받아줌
+      if (window.Kakao) {
+        const kakao = window.Kakao;
+        if (!kakao.isInitialized()) {
+          kakao.init(`${process.env.REACT_APP_KAKAO_JS_KEY}`); // 카카오에서 제공받은 javascript key를 넣어줌
+        }
 
-      kakao.Link.sendDefault({
-        // 링크 공유 중 여러가지 타입
-        objectType: 'feed',
-        content: {
-          title: modal.data.folderName,
-          description: modal.data.sharingUrl,
-          imageUrl: '이미지 url',
-          link: {
-            mobileWebUrl: modal.data.sharingUrl,
-            webUrl: modal.data.sharingUrl,
-          },
-        },
-        buttons: [
-          {
+        kakao.Link.sendDefault({
+          // 링크 공유 중 여러가지 타입
+          objectType: 'feed',
+          content: {
             title: modal.data.folderName,
+            description: modal.data.sharingUrl,
+            imageUrl: '이미지 url',
             link: {
               mobileWebUrl: modal.data.sharingUrl,
               webUrl: modal.data.sharingUrl,
             },
           },
-        ],
-      });
+          buttons: [
+            {
+              title: modal.data.folderName,
+              link: {
+                mobileWebUrl: modal.data.sharingUrl,
+                webUrl: modal.data.sharingUrl,
+              },
+            },
+          ],
+        });
+      }
     }
   };
